requirements.txt
datasets>=1.17.0
transformers==4.40.2
accelerate==0.30.1
bitsandbytes==0.43.1
loralib
peft>=0.3.0
tyro>=0.5.7
trl==0.9.4
databricks-sdk==0.27.1
langchain
langchain-community
mlflow==2.13.0
deepspeed==0.14.2
databricks-sql-connector
mosaicml-streaming
tiktoken